Changed texture sizes to more reasonable int

+++ Vocaluxe/Base/CDraw.cs
@@ -202,7 +202,7 @@ namespace Vocaluxe.Base
             RectangleF bounds = new RectangleF(staticBounds.Rect.X, staticBounds.Rect.Y, staticBounds.Rect.W, staticBounds.Rect.H);
             RectangleF rect;
 
-            CHelper.SetRect(bounds, out rect, texture.Width / texture.Height, aspect);
+            CHelper.SetRect(bounds, out rect, (float)texture.Width / texture.Height, aspect);
             DrawTexture(texture, new SRectF(rect.X, rect.Y, rect.Width, rect.Height, staticBounds.Rect.Z),
                         texture.Color, new SRectF(bounds.X, bounds.Y, bounds.Width, bounds.Height, 0f));
         }
